+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         >
-        <activity android:name="com.freshdigitable.dlfeey.MainActivity">
+        <activity android:name="com.freshdigitable.dlfeey.feed.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
