+++ app/src/main/java/com/freshdigitable/dlfeey/DlfeeyApp.kt
@@ -2,6 +2,7 @@ package com.freshdigitable.dlfeey
 
 import android.app.Activity
 import android.app.Application
+import com.freshdigitable.dlfeey.di.DaggerAppComponent
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.HasActivityInjector
