+++ app/src/main/java/com/freshdigitable/dlfeey/di/ActivityBuilder.kt
@@ -1,6 +1,8 @@
-package com.freshdigitable.dlfeey
+package com.freshdigitable.dlfeey.di
 
 import androidx.recyclerview.widget.RecyclerView
+import com.freshdigitable.dlfeey.feed.MainActivity
+import com.freshdigitable.dlfeey.feed.MainActivityModule
 import dagger.Module
 import dagger.Provides
 import dagger.android.ContributesAndroidInjector
