+++ app/src/main/java/com/freshdigitable/dlfeey/di/ActivityScoped.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.dlfeey
+package com.freshdigitable.dlfeey.di
 
 import javax.inject.Scope
