+++ app/src/main/java/com/freshdigitable/dlfeey/feed/FeedRepository.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.dlfeey
+package com.freshdigitable.dlfeey.feed
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
