+++ app/src/main/java/com/freshdigitable/dlfeey/feed/FeedViewModel.kt
@@ -1,4 +1,4 @@
-package com.freshdigitable.dlfeey
+package com.freshdigitable.dlfeey.feed
 
 import androidx.lifecycle.ViewModel
 import javax.inject.Inject
